[crypto] modularize union macros
